+++ src/components/organisms/FormLayout.js
@@ -1,5 +1,4 @@
-import React from "react";
-
+import React from 'react';
 import WideBody from "../atoms/WideBody";
 import BodyContainer from "../atoms/BodyContainer";
 import HeroImage from "../atoms/HeroImage";
@@ -11,17 +10,19 @@ const FormLayout = ({
   imageText,
   formHeader,
   formParagraph
-}) => (
-  <WideBody>
-    <BodyContainer>
-      <HeroImage src={imageType} alt={imageText} />
-      <Form
-        ctaText={ctaText}
-        formHeader={formHeader}
-        formParagraph={formParagraph}
-      />
-    </BodyContainer>
-  </WideBody>
-);
+}) => {
+  return (
+    <WideBody>
+      <BodyContainer>
+        <HeroImage src={imageType} alt={imageText} />
+        <Form
+          ctaText={ctaText}
+          formHeader={formHeader}
+          formParagraph={formParagraph}
+        />
+      </BodyContainer>
+    </WideBody>
+  );
+}
 
 export default FormLayout;
