+++ src/components/views/LoginPage.jsx
@@ -1,18 +1,8 @@
 import React from "react";
-import { useSelector } from "react-redux";
-import { Redirect, useLocation } from "react-router-dom";
-
 import image from "../../assets/Login.png";
 import { UserOnboarding } from "../templates";
 
 const LoginPage = () => {
-  const { token } = useSelector(state => state.currentUser);
-  const { state } = useLocation();
-
-  if (token) {
-    return <Redirect to={state?.from || '/dashboard'} />;
-  }
-
   return (
     <UserOnboarding
       ctaText="Log In"
